+++ starfish/__init__.py
@@ -1,18 +1,12 @@
-import logging
-from starfish.ocean.ocean import Ocean      # noqa: F401
-
 """
+
     Starfish-py module
 
-    This module contains the following classes
-
-    * :class:`Account` The Account class.
-    * :class:`Ocean` The main Ocean connection class.
-    * :data:`logger` The logger class.
 """
+
+from starfish.d_network import DNetwork     # noqa: F401
+from starfish.ocean.ocean import Ocean      # noqa: F401
+
 
 __author__ = """DEX.sg"""
 __version__ = '0.6.7'
-
-
-logger = logging.getLogger(__name__)