+++ starfish/account/__init__.py
@@ -1,3 +1,7 @@
+"""
 
-from starfish.account.account import Account                        # noqa: F401
-from starfish.account.starfish_account import StarfishAccount        # noqa: F401
+    Account module
+
+"""
+from starfish.account.account import Account                            # noqa: F401
+from starfish.account.starfish_account import StarfishAccount           # noqa: F401