+++ starfish/agent/__init__.py
@@ -1,6 +1,11 @@
+"""
+
+    Agent Module
 
 
-from starfish.agent.agent_base import AgentBase         # noqa: F401
-from starfish.agent.memory_agent import MemoryAgent     # noqa: F401
-from starfish.agent.remote_agent import RemoteAgent     # noqa: F401
-from starfish.agent.squid_agent import SquidAgent       # noqa: F401
+"""
+
+from starfish.agent.agent_base import AgentBase                         # noqa: F401
+from starfish.agent.memory_agent import MemoryAgent                     # noqa: F401
+from starfish.agent.remote_agent import RemoteAgent                     # noqa: F401
+from starfish.agent.squid_agent import SquidAgent                       # noqa: F401