+++ starfish/job/__init__.py
@@ -1,4 +1,8 @@
+"""
 
+    Job Module
 
-from starfish.job.job import Job                # noqa: F401
-from starfish.job.job_base import JobBase       # noqa: F401
+"""
+
+from starfish.job.job import Job                               # noqa: F401
+from starfish.job.job_base import JobBase                      # noqa: F401