+++ starfish/listing/__init__.py
@@ -1,3 +1,7 @@
+"""
 
-from starfish.listing.listing import Listing                # noqa: F401
-from starfish.listing.listing_base import ListingBase       # noqa: F401
+    Listing Module
+
+"""
+from starfish.listing.listing import Listing                    # noqa: F401
+from starfish.listing.listing_base import ListingBase           # noqa: F401