+++ starfish/metadata/__init__.py
@@ -1,3 +1,10 @@
-from starfish.metadata.metadata import Metadata     # noqa: F401
+"""
+
+    Metadata module
+
+
+"""
+
+from starfish.metadata.metadata import Metadata                                    # noqa: F401
 
 METADATA_PROVENANCE_NAME = 'provenance'