+++ tests/integration_dnetwork/contract/test_contract_manager.py
@@ -1,0 +1,13 @@
+"""
+
+
+    Test Contract Manager
+
+
+"""
+
+from starfish.contract import ContractManager
+
+def test_contract_load(dnetwork):
+    manager = ContractManager(dnetwork.web3, 'starfish.contract')
+    contract = manager.load('DirectPurchaseContract', dnetwork.network_name)