+++ tests/integration_dnetwork/contract/test_direct_purchase_contract.py
@@ -1,33 +1,24 @@
 import pytest
-
-
-from starfish.contract import ContractManager
+import secrets
 
 TOKEN_AMOUNT_TO_TRANSFER = 100
-REFERENCE_1 = 4
-REFERENCE_2 = 5
 
 
-"""
-
-
-
-"""
-def test_direct_purchase(config, starfish_accounts):
+def test_direct_purchase(dnetwork, starfish_accounts):
     """
 
     Transfer funds from the 'buy' account -> 'sell' account
 
     """
-    manager = ContractManager(config.keeper_url)
-    direct_contract = manager.load('DirectPurchaseContract')
-    ocean_token_contract = manager.load('OceanTokenContract')
-    dispenser_contract = manager.load('DispenserContract')
+    direct_contract = dnetwork.get_contract('DirectPurchase')
+    ocean_token_contract = dnetwork.get_contract('OceanToken')
+    dispenser_contract = dnetwork.get_contract('Dispenser')
 
 
-    from_account = starfish_accounts['purchaser']
-    to_account = starfish_accounts['publisher']
-
+    from_account = starfish_accounts[0]
+    to_account = starfish_accounts[1]
+    ref_1 = secrets.token_hex(32)
+    ref_2 = secrets.token_hex(32)
 
     tx_hash = dispenser_contract.request_tokens(TOKEN_AMOUNT_TO_TRANSFER, from_account)
     receipt = dispenser_contract.wait_for_receipt(tx_hash)
@@ -53,8 +44,8 @@ def test_direct_purchase(config, starfish_accounts):
         from_account,
         to_account.address,
         TOKEN_AMOUNT_TO_TRANSFER,
-        REFERENCE_1,
-        REFERENCE_2
+        ref_1,
+        ref_2
     )
     receipt = dispenser_contract.wait_for_receipt(tx_hash)
     # print('send token and log receipt', receipt)
@@ -69,16 +60,19 @@ def test_direct_purchase(config, starfish_accounts):
     assert(from_balance - TOKEN_AMOUNT_TO_TRANSFER == new_from_balance)
     assert(to_balance + TOKEN_AMOUNT_TO_TRANSFER == new_to_balance)
 
-def test_is_paid(config, starfish_accounts):
-    manager = ContractManager(config.keeper_url)
-    direct_contract = manager.load('DirectPurchaseContract')
-    from_account = starfish_accounts['purchaser']
-    to_account = starfish_accounts['publisher']
-
-    isPaid = direct_contract.check_is_paid(
+    is_sent = direct_contract.check_is_paid(
         from_account.address,
         to_account.address,
         TOKEN_AMOUNT_TO_TRANSFER,
-        REFERENCE_1
+        ref_1
     )
-    assert(isPaid)
+    assert(is_sent)
+
+    is_sent = direct_contract.check_is_paid(
+        from_account.address,
+        to_account.address,
+        TOKEN_AMOUNT_TO_TRANSFER,
+        ref_1,
+        ref_2
+    )
+    assert(is_sent)